A major redo.
The chiltepin executable was removed and replaced with a test.scala file.
The setup done by chiltepin was moved into a new wfgateway class.
A task class was created for building workflow tasks.
The workflow calss was reworked such that it is part of the DSL.
Implicits were reworked as needed.
The build.sbt was reworked to update to newer versions and rework assembly task.
Two new shell utilities created for compiling and running workflow scala codes.
Code was added to whoami to allow it to work in Darwin OS.